refactor: component type exports
