+++ packages/oruga-next/src/components/autocomplete/index.ts
@@ -4,9 +4,6 @@ import Autocomplete from "./Autocomplete.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export autocomplete specific types */
-export type * from "./types";
-
 /** export autocomplete plugin */
 export default {
     install(app: App) {
