+++ packages/oruga-next/src/components/button/index.ts
@@ -4,9 +4,6 @@ import Button from "./Button.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export button specific types */
-export type * from "./types";
-
 /** export button plugin */
 export default {
     install(app: App) {
