+++ packages/oruga-next/src/components/carousel/index.ts
@@ -5,9 +5,6 @@ import CarouselItem from "./CarouselItem.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export carousel specific types */
-export type * from "./types";
-
 /** export carousel plugin */
 export default {
     install(app: App) {
