+++ packages/oruga-next/src/components/checkbox/index.ts
@@ -4,9 +4,6 @@ import Checkbox from "./Checkbox.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export checkbox specific types */
-export type * from "./types";
-
 /** export checkbox plugin */
 export default {
     install(app: App) {
