+++ packages/oruga-next/src/components/datetimepicker/index.ts
@@ -4,9 +4,6 @@ import Datetimepicker from "./Datetimepicker.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export datetimepicker specific types */
-export type * from "./types";
-
 /** export datetimepicker plugin */
 export default {
     install(app: App) {
