+++ packages/oruga-next/src/components/field/index.ts
@@ -4,9 +4,6 @@ import Field from "./Field.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export field specific types */
-export type * from "./types";
-
 /** export field plugin */
 export default {
     install(app: App) {
