+++ packages/oruga-next/src/components/modal/types.ts
@@ -1,0 +1,5 @@
+import type { ComponentProps } from "vue-component-type-helpers";
+
+import Modal from "./Modal.vue";
+
+export type ModalProps = ComponentProps<typeof Modal>;
