+++ packages/oruga-next/src/components/notification/index.ts
@@ -9,10 +9,7 @@ import {
 } from "@/utils/plugins";
 
 /** export notification specific types */
-export type {
-    NotifcationNoticeProps,
-    NotifcationProps,
-} from "./NotificationProgrammatic";
+export type { NotifcationNoticeProps, NotifcationProps } from "./types";
 
 /** export notification plugin */
 export default {
