+++ packages/oruga-next/src/components/radio/index.ts
@@ -4,9 +4,6 @@ import Radio from "./Radio.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export radio specific types */
-export type * from "./types";
-
 /** export radio plugin */
 export default {
     install(app: App) {
