+++ packages/oruga-next/src/components/select/index.ts
@@ -5,7 +5,7 @@ import Select from "./Select.vue";
 import { registerComponent } from "@/utils/plugins";
 
 /** export select specific types */
-export type { OptionsItem } from "./utils";
+export type { OptionsItem } from "./types";
 
 /** export select plugin */
 export default {
