+++ packages/oruga-next/src/components/select/types.ts
@@ -2,11 +2,11 @@
  * Options should always be formatted as an array of objects with label and value
  * properties.
  */
-export interface OptionsItem<V = unknown> {
+export type OptionsItem<V = unknown> = {
     label: string;
     value: V;
     attrs?: {
         disabled?: boolean;
     } & Record<string, any>;
     [index: string]: any;
-}
+};
