+++ packages/oruga-next/src/components/sidebar/types.ts
@@ -1,0 +1,5 @@
+import type { ComponentProps } from "vue-component-type-helpers";
+
+import Sidebar from "./Sidebar.vue";
+
+export type SidebarProps = ComponentProps<typeof Sidebar>;
