+++ packages/oruga-next/src/components/skeleton/index.ts
@@ -4,9 +4,6 @@ import Skeleton from "./Skeleton.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export skeleton specific types */
-export type * from "./types";
-
 /** export skeleton plugin */
 export default {
     install(app: App) {
