+++ packages/oruga-next/src/components/slider/types.ts
