+++ packages/oruga-next/src/components/steps/types.ts
