+++ packages/oruga-next/src/components/switch/index.ts
@@ -4,9 +4,6 @@ import Switch from "./Switch.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export switch specific types */
-export type * from "./types";
-
 /** export switch plugin */
 export default {
     install(app: App) {
