+++ packages/oruga-next/src/components/table/types.ts
