+++ packages/oruga-next/src/components/tabs/types.ts
