+++ packages/oruga-next/src/components/taginput/index.ts
@@ -4,9 +4,6 @@ import Taginput from "./Taginput.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export taginput specific types */
-export type * from "./types";
-
 /** export taginput plugin */
 export default {
     install(Vue: App) {
