+++ packages/oruga-next/src/components/timepicker/index.ts
@@ -4,9 +4,6 @@ import Timepicker from "./Timepicker.vue";
 
 import { registerComponent } from "@/utils/plugins";
 
-/** export timepicker specific types */
-export type * from "./types";
-
 /** export timepicker plugin */
 export default {
     install(app: App) {
